Remove accidentally broken binding
In cleaning up the implementation of a bugfix, I accidentally left in a
binding that depends on a method that has been renamed, causing the
component to crash with an exception. The method is only used by method
reference and not passed around as a callback, so the binding can safely
be removed.